[dart2wasm] Use WasmI32 (instead of double) as representation of integers across interop boundary
Currently code uses `double` to represent array indices, array lengths
etc. This seems wrong because `double`s can loose precision. This is
also a high overhead.

Understandably one could use `double` and only start to loose precision
with more than 32 bits, but realistically string lenghts, array lengths
cannot exceed 31-bits (i.e. 2 GBs) in practice.

=> Use Wasmi32 to represent array indices, array lengths, bytes loaded
from integer arrays, ...

Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/372424
Commit-Queue: Martin Kustermann <kustermann@google.com>

+++ sdk/lib/_internal/wasm/lib/js_interop_patch.dart
@@ -215,7 +215,7 @@ extension ByteDataToJSDataView on ByteData {
     final t = this;
     return JSDataView._(JSValue(t is js_types.JSDataViewImpl
         ? t.toExternRef
-        : jsDataViewFromDartByteData(t, lengthInBytes.toDouble())));
+        : jsDataViewFromDartByteData(t, lengthInBytes)));
   }
 }
